add maxspeed=walk and assign it 5 kph
